trying to remove some of the horrible horrible cruf

+++ src/main/scala/puck/linalg/CLMatrix.scala
@@ -39,7 +39,7 @@ final class CLMatrix[@specialized(Int, Float, Double) V](val rows: Int,
                                                          val isTranspose: Boolean = false)(implicit val queue: CLQueue, ct: ClassTag[V])
   extends Matrix[V] with MatrixLike[V, CLMatrix[V]] {
   /** Creates a matrix with the specified data array, rows, and columns. */
-  def this(rows: Int, cols: Int)(implicit ct: ClassTag[V], context: CLContext, queue: CLQueue) = this(rows, cols, context.createBuffer[V](CLMem.Usage.InputOutput, ct.runtimeClass.asInstanceOf[Class[V]], rows * cols), 0, rows)
+  def this(rows: Int, cols: Int)(implicit ct: ClassTag[V], context: CLContext, queue: CLQueue) = this(rows, cols, context.createBuffer[V](CLMem.Usage.InputOutput, ct.runtimeClass.asInstanceOf[Class[V]], rows.toLong * cols), 0, rows)
   /** Creates a matrix with the specified data array, rows, and columns. Data must be column major */
   def this(rows: Int, cols: Int, data: CLBuffer[V], offset: Int = 0)(implicit queue: CLQueue, ct: ClassTag[V]) = this(rows, cols, data, offset, rows)
